% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle360.R
\name{angle360}
\alias{angle360}
\title{Calculate the angle (in radians) of the hypotenuse between vectors x and y relative to axis y.}
\usage{
angle360(x, y)
}
\arguments{
\item{x, y}{Numerical vectors.}
}
\value{
A vector of radian values.
}
\description{
Calculate the angle (in radians) of the hypotenuse between vectors x and y relative to axis y.
}
