% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degrees}
\alias{degrees}
\title{Convert radian angles to degrees.}
\usage{
degrees(a)
}
\arguments{
\item{a}{vector of angles in radians.}
}
\value{
A vector of angles in degrees.
}
\description{
Convert radian angles to degrees.
}
