% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_angle.R
\name{resample_angle}
\alias{resample_angle}
\title{Bootstrap vector norm and pleiotropic effects}
\usage{
resample_angle(x, y, SE_x, SE_y, ndraws = 1000, conf_interval = 0.95,
  width = 1/8)
}
\arguments{
\item{x}{Values along the \emph{x}- and \emph{y}-axes. These can be single
floats, vectors of floats, or matrices of floats.}

\item{y}{Values along the \emph{x}- and \emph{y}-axes. These can be single
floats, vectors of floats, or matrices of floats.}

\item{ndraws}{Integer number of bootstrap samples to draw.}

\item{conf_interval}{Probability coverage of the interval.}

\item{width}{Bin width for calculating the bootstrap distribution of \eqn{\tau}.}
}
\value{
A list of two elements giving confidence intervals for (z) vector
length and (tau) \eqn{\tau} values for each pair of elements in \code{x} and
\code{y}.
}
\description{
Parametric bootstrapping of vector length, and \eqn{\tau} assuming normality
of errors.
}
\details{
If \emph{x} and \emph{y} are estimated with normally-distributed errors
\eqn{\sigma_x} and \eqn{\sigma_y}, the joint bootstrap distribution can be
estimated by drawing pairs of values from \eqn{N(x, \sigma_x)} and
\eqn{N(y, \sigma_y)}. The length of each drawn vector and its \eqn{\tau} are
calculated, and confidence intervals drawn from their distribution.

In addition, the bootstrap distribution of \eqn{\tau} \emph{across} pairs of
observations in \emph{x} and \emph{y} is calculated by summing distributions
for each individual pair. Note that in contras to \code{tau_distribution}
this does incorporate confidence in position by summing \emph{q}; uncertainty
about distribution shape is accounted for through the bootstrapping process.
}
