% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SE_angle}
\alias{SE_angle}
\title{Standard error of an angle}
\usage{
SE_angle(x, y, SE_x, SE_y)
}
\arguments{
\item{x, y}{vectors of values for the \emph{x} and \emph{y} dimensions}

\item{SE_x, SE_y}{Vectors of standard errors for \emph{x} and \emph{y}}
}
\value{
Vector of standard errors for \eqn{\theta}.
}
\description{
Calculate the standard error of an acute angle in a right-angle triangle.
}
