% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SE_norm}
\alias{SE_norm}
\title{Standard error of the L2 norm of a vector.}
\usage{
SE_norm(x, y, SE_x, SE_y, norm)
}
\arguments{
\item{x, y}{vectors of values for the \emph{x} and \emph{y} dimensions}

\item{SE_x, SE_y}{Vectors of standard errors for \emph{x} and \emph{y}}

\item{norm}{Vector of L2 norms between \emph{x} and \emph{y} (hypotenuse between *x*
and *y*)}
}
\value{
Vector of norms for each pair of values in \emph{x} and \emph{y}
}
\description{
Calculate the standard error of the L2 norm for vectors of pairs of
coordinates.
}
