% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sintillate.R
\name{sintillate}
\alias{sintillate}
\title{Quantify pleiotropic effects on two variables.}
\usage{
sintillate(x, y)
}
\arguments{
\item{x, y}{Values along the \emph{x}- and \emph{y}-axes. These can be single
floats, vectors of floats, or matrices of floats.}
}
\value{
Data frame including input data, vector norm angle, and \eqn{\tau}.
If matrices were supplied, this returns a list of these data.
}
\description{
\code{sintillate} calculates vector length, angle \eqn{\tau} for vectors
\emph{x} and \emph{y}.
}
\details{
Calculate basic summary statistics about a pair of vectors.
This can also be used to calculate bootstrap values for each parameter by
providing matrices rather than vectors of input values.
}
