% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\title{Quantify tau between two variables.}
\usage{
tau(a)
}
\arguments{
\item{a}{Vector of angles in radians.}
}
\value{
Vector of \eqn{\tau} values between -1 and 1.
}
\description{
\code{tau} calculates the statistic \eqn{\tau} between two data points.
}
