% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_distribution.R
\name{tau_distribution}
\alias{tau_distribution}
\title{Distribution of pleiotropic effects}
\usage{
tau_distribution(tau, q, width = 1/8)
}
\arguments{
\item{tau}{Vector of \eqn{\tau} values.}

\item{width}{Width of each bin. Passed to \code{cut}. Since \eqn{\tau} can
only be between -1 and 1, it is meaningless to assign a width greater than
that interval.}

\item{tau}{Vector of \eqn{q} values.}
}
\value{
Data.frame listing bining levels, the midpoints for each bin, and the
sums of \emph{q}-values in each bin.
}
\description{
Bins values of \eqn{\tau} across pairs of observations so that a histogram can be created.
}
\details{
\code{tau_distribution} groups values of \eqn{tau} over observations into bins
defined by \code{width}. Within each bin, values for \emph{q} for each
\eqn{\tau} are summed. The resulting sum can be interpreted as the
probability of observing a vector with \eqn{\tau} in that bin, accoutning for
the robustness of the estimate of \eqn{\tau}.
}
