% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_norm}
\alias{vector_norm}
\title{Calculate the L2 norm of a 2-d vector}
\usage{
vector_norm(x, y)
}
\arguments{
\item{x, y}{vectors of values for the \emph{x} and \emph{y} dimensions}
}
\value{
Vector of norms for each pair of values in \emph{x} and \emph{y}
}
\description{
Calculate the L2 norm of a 2-d vector
}
