% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sintillate.R
\name{sintillate}
\alias{sintillate}
\title{Quantify pleiotropic effects on two variables.}
\usage{
sintillate(x, y, SE_x = NULL, SE_y = NULL)
}
\arguments{
\item{x}{Value, or vector of values allongthe \emph{x}-axis.}

\item{y}{Value, or vector of values allongthe \emph{y}-axis.}

\item{SE_x, SE_y}{Optional standard errors for elements in \emph{x} and \emph{y}.}
}
\value{
Data frame including input data, vector norm angle, and \eqn{\tau}.
If standard errors are supplied, standard errors for the norm, angle, are
provided, as well as \emph{q} assuming normality of errors..
}
\description{
\code{sintillate} calculates \eqn{\tau} and \eqn{q} for vectors \emph{x} and \emph{y}.
}
